Split conflict-flow autosave tests into their own suite
Main removed this test file's max-lines disable (#7548); merged with the
new conflict tests it exceeded the cap. The controller suite returns to
main's exact version; the changed-on-disk tests move to
editor-autosave-conflict-flow.test.ts on a shared window/store fixture.
