+++ src/renderer/src/components/editor/editor-autosave-controller.test.ts
@@ -1,5 +1,3 @@
-/* eslint-disable max-lines -- Why: autosave behavior depends on event wiring,
-   dirty drafts, remote routing, quiesce, and failure cleanup in one harness. */
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest'
 import { createStore, type StoreApi } from 'zustand/vanilla'
 import { createEditorSlice } from '@/store/slices/editor'
@@ -8,18 +6,10 @@ import {
   ORCA_EDITOR_PREPARE_HOT_EXIT_EVENT,
   ORCA_EDITOR_SAVE_DIRTY_FILES_EVENT
 } from '../../../../shared/editor-save-events'
-import {
-  ORCA_EDITOR_EXTERNAL_FILE_CHANGE_EVENT,
-  requestEditorFileSave,
-  requestEditorSaveQuiesce
-} from './editor-autosave'
+import { requestEditorFileSave, requestEditorSaveQuiesce } from './editor-autosave'
 import { attachEditorAutosaveController } from './editor-autosave-controller'
 import { registerPendingEditorFlush } from './editor-pending-flush'
-import {
-  __clearSelfWriteRegistryForTests,
-  hasRecentSelfWrite,
-  recordSelfWrite
-} from './editor-self-write-registry'
+import { __clearSelfWriteRegistryForTests, hasRecentSelfWrite } from './editor-self-write-registry'
 import {
   createCompatibleRuntimeStatusResponseIfNeeded,
   type RuntimeEnvironmentCallRequest
@@ -624,282 +614,6 @@ describe('attachEditorAutosaveController', () => {
     try {
       await expect(requestEditorFileSave({ fileId: '/repo/file.md' })).rejects.toThrow('disk full')
       expect(hasRecentSelfWrite('/repo/file.md')).toBe(false)
-    } finally {
-      cleanup()
-    }
-  })
-
-  it('preserves dirty drafts and marks the tab changed-on-disk on external file change', () => {
-    const writeFile = vi.fn().mockResolvedValue(undefined)
-    const eventTarget = new EventTarget()
-    vi.stubGlobal('window', {
-      addEventListener: eventTarget.addEventListener.bind(eventTarget),
-      removeEventListener: eventTarget.removeEventListener.bind(eventTarget),
-      dispatchEvent: eventTarget.dispatchEvent.bind(eventTarget),
-      setTimeout: globalThis.setTimeout.bind(globalThis),
-      clearTimeout: globalThis.clearTimeout.bind(globalThis),
-      api: {
-        fs: {
-          writeFile
-        }
-      }
-    } satisfies WindowStub)
-
-    const store = createEditorStore()
-    store.getState().openFile({
-      filePath: '/repo/file.ts',
-      relativePath: 'file.ts',
-      worktreeId: 'wt-1',
-      language: 'typescript',
-      mode: 'edit'
-    })
-    store.getState().setEditorDraft('/repo/file.ts', 'unsaved edit')
-    store.getState().markFileDirty('/repo/file.ts', true)
-
-    const cleanup = attachEditorAutosaveController(store)
-    try {
-      window.dispatchEvent(
-        new CustomEvent(ORCA_EDITOR_EXTERNAL_FILE_CHANGE_EVENT, {
-          detail: { worktreeId: 'wt-1', worktreePath: '/repo', relativePath: 'file.ts' }
-        })
-      )
-
-      const file = store.getState().openFiles[0]
-      expect(file?.isDirty).toBe(true)
-      expect(file?.externalMutation).toBe('changed')
-      expect(store.getState().editorDrafts['/repo/file.ts']).toBe('unsaved edit')
-    } finally {
-      cleanup()
-    }
-  })
-
-  it('clears drafts and a stale changed-on-disk mark for clean tabs on external file change', () => {
-    const writeFile = vi.fn().mockResolvedValue(undefined)
-    const eventTarget = new EventTarget()
-    vi.stubGlobal('window', {
-      addEventListener: eventTarget.addEventListener.bind(eventTarget),
-      removeEventListener: eventTarget.removeEventListener.bind(eventTarget),
-      dispatchEvent: eventTarget.dispatchEvent.bind(eventTarget),
-      setTimeout: globalThis.setTimeout.bind(globalThis),
-      clearTimeout: globalThis.clearTimeout.bind(globalThis),
-      api: {
-        fs: {
-          writeFile
-        }
-      }
-    } satisfies WindowStub)
-
-    const store = createEditorStore()
-    store.getState().openFile({
-      filePath: '/repo/file.ts',
-      relativePath: 'file.ts',
-      worktreeId: 'wt-1',
-      language: 'typescript',
-      mode: 'edit'
-    })
-    store.getState().setExternalMutation('/repo/file.ts', 'changed')
-
-    const cleanup = attachEditorAutosaveController(store)
-    try {
-      window.dispatchEvent(
-        new CustomEvent(ORCA_EDITOR_EXTERNAL_FILE_CHANGE_EVENT, {
-          detail: { worktreeId: 'wt-1', worktreePath: '/repo', relativePath: 'file.ts' }
-        })
-      )
-
-      const file = store.getState().openFiles[0]
-      expect(file?.isDirty).toBe(false)
-      expect(file?.externalMutation).toBeUndefined()
-    } finally {
-      cleanup()
-    }
-  })
-
-  it('does not backstop-mark a dirty tab for the echo of its own save', () => {
-    const writeFile = vi.fn().mockResolvedValue(undefined)
-    const eventTarget = new EventTarget()
-    vi.stubGlobal('window', {
-      addEventListener: eventTarget.addEventListener.bind(eventTarget),
-      removeEventListener: eventTarget.removeEventListener.bind(eventTarget),
-      dispatchEvent: eventTarget.dispatchEvent.bind(eventTarget),
-      setTimeout: globalThis.setTimeout.bind(globalThis),
-      clearTimeout: globalThis.clearTimeout.bind(globalThis),
-      api: {
-        fs: {
-          writeFile
-        }
-      }
-    } satisfies WindowStub)
-
-    const store = createEditorStore()
-    store.getState().openFile({
-      filePath: '/repo/file.ts',
-      relativePath: 'file.ts',
-      worktreeId: 'wt-1',
-      language: 'typescript',
-      mode: 'edit'
-    })
-    store.getState().setEditorDraft('/repo/file.ts', 'typed during save')
-    store.getState().markFileDirty('/repo/file.ts', true)
-    // Why: the combined-Changes reload notification routes through the
-    // controller for the saved path — a fresh self-write stamp means the
-    // event is Orca's own echo, not an external change.
-    recordSelfWrite('/repo/file.ts', 'orca save')
-
-    const cleanup = attachEditorAutosaveController(store)
-    try {
-      window.dispatchEvent(
-        new CustomEvent(ORCA_EDITOR_EXTERNAL_FILE_CHANGE_EVENT, {
-          detail: { worktreeId: 'wt-1', worktreePath: '/repo', relativePath: 'file.ts' }
-        })
-      )
-
-      const file = store.getState().openFiles[0]
-      expect(file?.externalMutation).toBeUndefined()
-      expect(file?.isDirty).toBe(true)
-      expect(store.getState().editorDrafts['/repo/file.ts']).toBe('typed during save')
-    } finally {
-      cleanup()
-    }
-  })
-
-  it('suspends autosave while a tab is marked changed-on-disk and resumes when cleared', async () => {
-    const writeFile = vi.fn().mockResolvedValue(undefined)
-    const eventTarget = new EventTarget()
-    vi.stubGlobal('window', {
-      addEventListener: eventTarget.addEventListener.bind(eventTarget),
-      removeEventListener: eventTarget.removeEventListener.bind(eventTarget),
-      dispatchEvent: eventTarget.dispatchEvent.bind(eventTarget),
-      setTimeout: globalThis.setTimeout.bind(globalThis),
-      clearTimeout: globalThis.clearTimeout.bind(globalThis),
-      api: {
-        fs: {
-          writeFile
-        }
-      }
-    } satisfies WindowStub)
-
-    const store = createEditorStore()
-    store.getState().openFile({
-      filePath: '/repo/file.ts',
-      relativePath: 'file.ts',
-      worktreeId: 'wt-1',
-      language: 'typescript',
-      mode: 'edit'
-    })
-    store.getState().setEditorDraft('/repo/file.ts', 'user edit')
-    store.getState().markFileDirty('/repo/file.ts', true)
-    store.getState().setExternalMutation('/repo/file.ts', 'changed')
-
-    const cleanup = attachEditorAutosaveController(store)
-    try {
-      await vi.advanceTimersByTimeAsync(1500)
-      expect(writeFile).not.toHaveBeenCalled()
-
-      // Keep My Edits clears the mark — autosave resumes and overwrites.
-      store.getState().setExternalMutation('/repo/file.ts', null)
-      await vi.advanceTimersByTimeAsync(1500)
-      expect(writeFile).toHaveBeenCalledWith({
-        filePath: '/repo/file.ts',
-        content: 'user edit'
-      })
-    } finally {
-      cleanup()
-    }
-  })
-
-  it('suspends autosave while a restored tab awaits disk baseline verification', async () => {
-    const writeFile = vi.fn().mockResolvedValue(undefined)
-    const eventTarget = new EventTarget()
-    vi.stubGlobal('window', {
-      addEventListener: eventTarget.addEventListener.bind(eventTarget),
-      removeEventListener: eventTarget.removeEventListener.bind(eventTarget),
-      dispatchEvent: eventTarget.dispatchEvent.bind(eventTarget),
-      setTimeout: globalThis.setTimeout.bind(globalThis),
-      clearTimeout: globalThis.clearTimeout.bind(globalThis),
-      api: {
-        fs: {
-          writeFile
-        }
-      }
-    } satisfies WindowStub)
-
-    const store = createEditorStore()
-    store.getState().openFile({
-      filePath: '/repo/file.ts',
-      relativePath: 'file.ts',
-      worktreeId: 'wt-1',
-      language: 'typescript',
-      mode: 'edit'
-    })
-    store.getState().setEditorDraft('/repo/file.ts', 'restored draft')
-    store.getState().markFileDirty('/repo/file.ts', true)
-    // Why: mimic hydration — the scan has not yet compared disk against the
-    // persisted baseline, so autosave must hold off (a slow remote read must
-    // not lose a race to this timer).
-    store.setState({
-      openFiles: store
-        .getState()
-        .openFiles.map((f) =>
-          f.id === '/repo/file.ts' ? { ...f, pendingDiskBaselineVerification: true } : f
-        )
-    } as never)
-
-    const cleanup = attachEditorAutosaveController(store)
-    try {
-      await vi.advanceTimersByTimeAsync(1500)
-      expect(writeFile).not.toHaveBeenCalled()
-
-      store.getState().clearPendingDiskBaselineVerification('/repo/file.ts')
-      await vi.advanceTimersByTimeAsync(1500)
-      expect(writeFile).toHaveBeenCalledWith({
-        filePath: '/repo/file.ts',
-        content: 'restored draft'
-      })
-    } finally {
-      cleanup()
-    }
-  })
-
-  it('clears the changed-on-disk mark after a successful save', async () => {
-    const writeFile = vi.fn().mockResolvedValue(undefined)
-    const eventTarget = new EventTarget()
-    vi.stubGlobal('window', {
-      addEventListener: eventTarget.addEventListener.bind(eventTarget),
-      removeEventListener: eventTarget.removeEventListener.bind(eventTarget),
-      dispatchEvent: eventTarget.dispatchEvent.bind(eventTarget),
-      setTimeout: globalThis.setTimeout.bind(globalThis),
-      clearTimeout: globalThis.clearTimeout.bind(globalThis),
-      api: {
-        fs: {
-          writeFile
-        }
-      }
-    } satisfies WindowStub)
-
-    const store = createEditorStore()
-    store.getState().openFile({
-      filePath: '/repo/file.ts',
-      relativePath: 'file.ts',
-      worktreeId: 'wt-1',
-      language: 'typescript',
-      mode: 'edit'
-    })
-    store.getState().setEditorDraft('/repo/file.ts', 'user version')
-    store.getState().markFileDirty('/repo/file.ts', true)
-    store.getState().setExternalMutation('/repo/file.ts', 'changed')
-
-    const cleanup = attachEditorAutosaveController(store)
-    try {
-      await requestEditorFileSave({ fileId: '/repo/file.ts' })
-
-      expect(writeFile).toHaveBeenCalledWith({
-        filePath: '/repo/file.ts',
-        content: 'user version'
-      })
-      const file = store.getState().openFiles[0]
-      expect(file?.isDirty).toBe(false)
-      expect(file?.externalMutation).toBeUndefined()
     } finally {
       cleanup()
     }
